move eigen of bfloat16
